+++ src/main/kotlin/csense/idea/kotlin/checked/exceptions/quickfixes/AddLambdaToCallthoughQuickFix.kt
@@ -1,10 +1,10 @@
 package csense.idea.kotlin.checked.exceptions.quickfixes
 
-//import csense.idea.kotlin.checked.exceptions.callthough.*
 import com.intellij.openapi.project.*
 import com.intellij.psi.*
 import csense.idea.base.bll.psi.*
 import csense.idea.base.bll.quickfixes.*
+import csense.idea.kotlin.checked.exceptions.bll.*
 import csense.idea.kotlin.checked.exceptions.bll.callthough.*
 import org.jetbrains.kotlin.psi.*
 
@@ -13,7 +13,7 @@ class AddLambdaToCallthoughQuickFix(
     private val parameterName: String
 ) : LocalQuickFixOnSingleKtElement<KtFunction>(lambdaFunction) {
     override fun getFamilyName(): String {
-        return "Csense - checked exceptions - add to callthough file quick fix"
+        return "${Constants.groupName} - add to callthough file quick fix"
     }
 
     override fun getText(): String {