+++ src/main/kotlin/csense/idea/kotlin/checked/exceptions/quickfixes/selectors/KtThrowExpressionQuickFixSelector.kt
@@ -8,7 +8,7 @@ import csense.kotlin.extensions.collections.list.*
 import org.jetbrains.kotlin.psi.*
 
 class KtThrowExpressionQuickFixSelector(
-    private val throwExpression: KtThrowExpression,
+    throwExpression: KtThrowExpression,
     uncaughtExceptions: List<KtPsiClass>
 ) : AbstractUncaughtExceptionQuickfixSelector(
     element = throwExpression,