+++ changelog.md
@@ -3,9 +3,9 @@
   - Most features are now working across android studio & / intellij and across various versions
   - Performance should be superb
   - Quick fixes should generally work pretty well now
-  - The settings page have now been reworked
   - a lot of things that previously did not work (or only partially worked) now works
-  - 
+  - Highlights simple Throws in kotlin doc
+  - should respect imports now
 
 # 1.2.0
 - Completely fixed a lot of bugs (via testing)